pyproject.toml
[tool.setuptools]
packages = ["app", "test"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "ecommerce"
authors = [{name = "Tien Pham", email = "dinhlang86@gmail.com"}]
description = "Ecommerce backend api using FastAPI"
requires-python = ">=3.8"
dynamic = ["version"]
dependencies = [
    "fastapi==0.111.0",
    "sqlmodel==0.0.18",
    "alembic==1.13.1",
    "asyncpg==0.29.0",
    "passlib==1.7.4",
    "python-jose==3.3.0",
    "cryptography==42.0.7",
    "bcrypt==4.1.3"
]

[project.optional-dependencies]
test = ["pytest==8.2.0",
    "pytest_asyncio==0.23.6",
    "SQLAlchemy-Utils==0.41.2",
    "types-sqlalchemy-utils==1.1.0"]

dev = ["black==24.4.2",
    "flake8==7.0.0", 
    "mypy==1.10.0", 
    "isort==5.13.2",
    "types-passlib==1.7.7.20240327",
    "types-python-jose==3.3.4.20240106"]

[tool.black]
line-length = 100
exclude = '''
/(
    \.git
  | \.mypy_cache
  | venv
  | \.venv
)/
'''

[tool.isort]
# make it compatible with black
profile = "black" 

[tool.mypy]
strict=true
implicit_reexport=true
exclude=[".venv"]

[tool.pytest.ini_options]
pythonpaths = ["./"]
asyncio_mode = "auto"
